+++ cosmos/lcd.go
@@ -180,14 +180,18 @@ func (lcd *LCD) getAccount() (*auth.BaseAccount, error) {
 		}
 		lcd.acc = auth.NewBaseAccount(accKb.GetAddress(), nil, accKb.GetPubKey(), 0, 0)
 	}
-	localSeq := lcd.acc.GetSequence()
-	if err := lcd.Get("auth/accounts/"+lcd.acc.GetAddress().String(), &lcd.acc); err != nil {
+	var accR *auth.BaseAccount
+	if err := lcd.Get("auth/accounts/"+lcd.acc.GetAddress().String(), &accR); err != nil {
 		return nil, err
 	}
-	// replace seq if sup
-	if localSeq > lcd.acc.GetSequence() {
-		lcd.acc.SetSequence(localSeq)
+	if accR.Address.Empty() {
+		return nil, fmt.Errorf("account %q doesn't exist", lcd.acc.GetAddress().String())
 	}
+	// replace seq if sup
+	if lcd.acc.GetSequence() > accR.GetSequence() {
+		accR.SetSequence(lcd.acc.GetSequence())
+	}
+	lcd.acc = accR
 	return lcd.acc, nil
 }