+++ e2e/orchestrator_nested_data_test.go
@@ -55,7 +55,6 @@ func testOrchestratorNestedData(runnerHash, instanceHash hash.Hash) func(t *test
 		})
 		t.Run("trigger process", func(t *testing.T) {
 			req := orchestrator.ExecutionCreateRequest{
-				Price:        "10000atto",
 				TaskKey:      "task_complex_trigger",
 				ExecutorHash: runnerHash,
 				Inputs: &types.Struct{